+++ src/settings/note-types-settings.js
@@ -6,8 +6,12 @@ import {
 } from 'react-intl';
 import { get } from 'lodash';
 
-import { TitleManager } from '@folio/stripes/core';
+import {
+  TitleManager,
+  useCallout,
+} from '@folio/stripes/core';
 import { ControlledVocab } from '@folio/stripes/smart-components';
+
 import { validate } from '../util';
 
 const propTypes = {
@@ -19,6 +23,7 @@ const propTypes = {
 
 const NoteTypesSettings = ({ stripes }) => {
   const { formatMessage } = useIntl();
+  const callout = useCallout();
   const ConnectedControlledVocab = stripes.connect(ControlledVocab);
 
   const paneTitle = formatMessage({ id: 'ui-notes.settings.noteTypes' });
@@ -31,6 +36,24 @@ const NoteTypesSettings = ({ stripes }) => {
 
   const suppressDelete = noteType => {
     return !canEdit || get(noteType, 'usage.isAssigned');
+  };
+
+  const handleCreateFail = (res) => {
+    res.json().then(body => {
+      const error = body?.errors?.[0];
+
+      if (!error) {
+        return;
+      }
+
+      if (error.code === 'NOTE_TYPES_LIMIT_REACHED') {
+        const limit = error.parameters.find(param => param.key === 'limit');
+        callout.sendCallout({
+          type: 'error',
+          message: formatMessage({ id: 'ui-notes.settings.maxAmount' }, { amount: limit?.value }),
+        });
+      }
+    });
   };
 
   const suppressEdit = () => !canEdit;
@@ -59,6 +82,7 @@ const NoteTypesSettings = ({ stripes }) => {
           name: label
         }}
         canCreate={canEdit}
+        onCreateFail={handleCreateFail}
         nameKey="name"
         id="noteTypes"
         sortby="name"